pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "fastmcp>=2.12.4",
    "google-genai>=1.43.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
